zld: move contents of Zld into MachO module
